feat: adding Lab Table Commands, and refactoring these Commands

+++ Program.cs
@@ -7,6 +7,7 @@ var databaseConfig = new DatabaseConfig();
 
 var databaseSetup = new DatabaseSetup(databaseConfig);
 var computerRepository = new ComputerRepository(databaseConfig);
+var labRepository = new LabRepository(databaseConfig);
 
 // Routing
 var modelName = args[0];
@@ -47,7 +48,15 @@ if(modelName == "Computer")
     if(modelAction == "Delete")
     {
         var id = Convert.ToInt32(args[2]);
-        computerRepository.Delete(id);
+
+        if (computerRepository.ExistsByID(id))
+        {
+            computerRepository.Delete(id);
+        }
+        else
+        {
+            Console.WriteLine($"O Computador {id} não existe.");
+        }
     }
 
     if(modelAction == "Show") 
@@ -71,44 +80,60 @@ if(modelName == "Lab")
     if(modelAction == "List")
     {
         Console.WriteLine("Lab List: ");
-        var connection = new SqliteConnection(databaseConfig.ConnectionString);
-        connection.Open();
-
-        var command = connection.CreateCommand();
-        command.CommandText = "SELECT * FROM Lab;";
-
-        var reader = command.ExecuteReader();
-
-        while (reader.Read())
+        foreach (var lab in labRepository.GetAll())
         {
-            Console.WriteLine(
-                $"{reader.GetInt32(0)}, {reader.GetInt32(1)}, {reader.GetString(2)}, {reader.GetChar(3)}"
-            );
-            
+            Console.WriteLine($"{lab.Id}, {lab.Number}, {lab.Name}, {lab.Block}");
         }
-
-        connection.Close();
     } 
     
     if (modelAction == "New")
     {
         var id = Convert.ToInt32(args[2]);
-        var number = args[3];
+        var number = Convert.ToInt32(args[3]);
         var name = args[4];
-        var block = args[5];
-        
-        var connection = new SqliteConnection(databaseConfig.ConnectionString);
-        connection.Open();
+        var block = Convert.ToChar(args[5]);
 
-        var command = connection.CreateCommand();
-        command.CommandText = "INSERT INTO Lab VALUES($id, $number, $name, $block)";
-        command.Parameters.AddWithValue("$id", id);
-        command.Parameters.AddWithValue("$number", number);
-        command.Parameters.AddWithValue("$name", name);
-        command.Parameters.AddWithValue("$block", block);
+        var lab = new Lab(id, number, name, block);
+        labRepository.Save(lab);
+    }  
 
-        command.ExecuteNonQuery();
+    if (modelAction == "Update")
+    {
+        var id = Convert.ToInt32(args[2]);
+        var number = Convert.ToInt32(args[3]);
+        var name = args[4];
+        var block = Convert.ToChar(args[5]);
 
-        connection.Close();
-    }   
+        var lab = new Lab(id, number, name, block);
+        labRepository.Update(lab);
+    }
+
+    if(modelAction == "Delete")
+    {
+        var id = Convert.ToInt32(args[2]);
+
+        if (labRepository.ExistsByID(id))
+        {
+            labRepository.Delete(id);
+        }
+        else
+        {
+            Console.WriteLine($"O Laboratório {id} não existe.");
+        }
+    }
+
+    if(modelAction == "Show") 
+    {
+        var id = Convert.ToInt32(args[2]);
+
+        if (labRepository.ExistsByID(id))
+        {
+            var lab = labRepository.GetById(id);
+            Console.WriteLine($"{lab.Id}, {lab.Number}, {lab.Name}, {lab.Block}");  
+        }
+        else
+        {
+            Console.WriteLine($"O Laboratório {id} não existe.");
+        }
+    } 
 }
